Move database fetchers to discovery service
