+++ lib/srv/db/common/errors.go
@@ -237,12 +237,3 @@ agent's service principal. See: https://goteleport.com/docs/database-access/guid
 		return trace.Wrap(err)
 	}
 }
-
-// IsUnrecognizedAWSEngineNameError checks if the err is non-nil and came from using an engine filter that the
-// AWS region does not recognize.
-func IsUnrecognizedAWSEngineNameError(err error) bool {
-	if err == nil {
-		return false
-	}
-	return strings.Contains(strings.ToLower(err.Error()), "unrecognized engine name")
-}
