+++ lib/srv/db/watcher.go
@@ -20,11 +20,13 @@ import (
 	"context"
 
 	"github.com/gravitational/trace"
+	"github.com/sirupsen/logrus"
 
 	"github.com/gravitational/teleport"
 	"github.com/gravitational/teleport/api/types"
 	"github.com/gravitational/teleport/lib/services"
-	"github.com/gravitational/teleport/lib/srv/db/cloud/watchers"
+	discovery "github.com/gravitational/teleport/lib/srv/discovery/common"
+	dbfetchers "github.com/gravitational/teleport/lib/srv/discovery/fetchers/db"
 )
 
 // startReconciler starts reconciler that registers/unregisters proxied
@@ -102,10 +104,18 @@ func (s *Server) startResourceWatcher(ctx context.Context) (*services.DatabaseWa
 // startCloudWatcher starts fetching cloud databases according to the
 // selectors and register/unregister them appropriately.
 func (s *Server) startCloudWatcher(ctx context.Context) error {
-	watcher, err := watchers.NewWatcher(ctx, watchers.WatcherConfig{
-		AWSMatchers:   s.cfg.AWSMatchers,
-		AzureMatchers: s.cfg.AzureMatchers,
-		Clients:       s.cfg.CloudClients,
+	awsFetchers, err := dbfetchers.MakeAWSFetchers(s.cfg.CloudClients, s.cfg.AWSMatchers)
+	if err != nil {
+		return trace.Wrap(err)
+	}
+	azureFechters, err := dbfetchers.MakeAzureFetchers(s.cfg.CloudClients, s.cfg.AzureMatchers)
+	if err != nil {
+		return trace.Wrap(err)
+	}
+
+	watcher, err := discovery.NewWatcher(ctx, discovery.WatcherConfig{
+		Fetchers: append(awsFetchers, azureFechters...),
+		Log:      logrus.WithField(trace.Component, "watcher:cloud"),
 	})
 	if err != nil {
 		if trace.IsNotFound(err) {
@@ -119,8 +129,13 @@ func (s *Server) startCloudWatcher(ctx context.Context) error {
 		defer s.log.Debug("Cloud database watcher done.")
 		for {
 			select {
-			case databases := <-watcher.DatabasesC():
-				s.monitoredDatabases.setCloud(databases)
+			case resources := <-watcher.ResourcesC():
+				databases, err := resources.AsDatabases()
+				if err == nil {
+					s.monitoredDatabases.setCloud(databases)
+				} else {
+					s.log.WithError(err).Warnf("Failed to convert resources to databases.")
+				}
 				select {
 				case s.reconcileCh <- struct{}{}:
 				case <-ctx.Done():
