+++ lib/srv/discovery/common/interfaces.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package fetchers
+package common
 
 import (
 	"context"
