+++ lib/srv/discovery/fetchers/aks.go
@@ -18,6 +18,7 @@ package fetchers
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/gravitational/trace"
 	"github.com/sirupsen/logrus"
@@ -26,6 +27,7 @@ import (
 	"github.com/gravitational/teleport/api/types"
 	"github.com/gravitational/teleport/lib/cloud/azure"
 	"github.com/gravitational/teleport/lib/services"
+	"github.com/gravitational/teleport/lib/srv/discovery/common"
 )
 
 type aksFetcher struct {
@@ -66,7 +68,7 @@ func (c *AKSFetcherConfig) CheckAndSetDefaults() error {
 }
 
 // NewAKSFetcher creates a new AKS fetcher configuration.
-func NewAKSFetcher(cfg AKSFetcherConfig) (Fetcher, error) {
+func NewAKSFetcher(cfg AKSFetcherConfig) (common.Fetcher, error) {
 	if err := cfg.CheckAndSetDefaults(); err != nil {
 		return nil, trace.Wrap(err)
 	}
@@ -136,4 +138,9 @@ func (a *aksFetcher) ResourceType() string {
 
 func (a *aksFetcher) Cloud() string {
 	return types.CloudAzure
+}
+
+func (a *aksFetcher) String() string {
+	return fmt.Sprintf("aksFetcher(ResourceGroups=%v, Regions=%v, Labels=%v)",
+		a.ResourceGroups, a.Regions, a.FilterLabels)
 }
