+++ lib/srv/discovery/fetchers/db/aws_rds.go
@@ -14,11 +14,12 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package watchers
+package db
 
 import (
 	"context"
 	"fmt"
+	"strings"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/rds"
@@ -27,12 +28,13 @@ import (
 	"github.com/sirupsen/logrus"
 
 	"github.com/gravitational/teleport/api/types"
+	libcloudaws "github.com/gravitational/teleport/lib/cloud/aws"
 	"github.com/gravitational/teleport/lib/services"
-	"github.com/gravitational/teleport/lib/srv/db/common"
+	"github.com/gravitational/teleport/lib/srv/discovery/common"
 )
 
-// rdsFetcherConfig is the RDS databases fetcher configuration.
-type rdsFetcherConfig struct {
+// RDSFetcherConfig is the RDS databases fetcher configuration.
+type RDSFetcherConfig struct {
 	// Labels is a selector to match cloud databases.
 	Labels types.Labels
 	// RDS is the RDS API client.
@@ -42,7 +44,7 @@ type rdsFetcherConfig struct {
 }
 
 // CheckAndSetDefaults validates the config and sets defaults.
-func (c *rdsFetcherConfig) CheckAndSetDefaults() error {
+func (c *RDSFetcherConfig) CheckAndSetDefaults() error {
 	if len(c.Labels) == 0 {
 		return trace.BadParameter("missing parameter Labels")
 	}
@@ -57,12 +59,14 @@ func (c *rdsFetcherConfig) CheckAndSetDefaults() error {
 
 // rdsDBInstancesFetcher retrieves RDS DB instances.
 type rdsDBInstancesFetcher struct {
-	cfg rdsFetcherConfig
+	awsFetcher
+
+	cfg RDSFetcherConfig
 	log logrus.FieldLogger
 }
 
-// newRDSDBInstancesFetcher returns a new RDS DB instances fetcher instance.
-func newRDSDBInstancesFetcher(config rdsFetcherConfig) (Fetcher, error) {
+// NewRDSDBInstancesFetcher returns a new RDS DB instances fetcher instance.
+func NewRDSDBInstancesFetcher(config RDSFetcherConfig) (common.Fetcher, error) {
 	if err := config.CheckAndSetDefaults(); err != nil {
 		return nil, trace.Wrap(err)
 	}
@@ -77,20 +81,20 @@ func newRDSDBInstancesFetcher(config rdsFetcherConfig) (Fetcher, error) {
 }
 
 // Get returns RDS DB instances matching the watcher's selectors.
-func (f *rdsDBInstancesFetcher) Get(ctx context.Context) (types.Databases, error) {
+func (f *rdsDBInstancesFetcher) Get(ctx context.Context) (types.ResourcesWithLabels, error) {
 	rdsDatabases, err := f.getRDSDatabases(ctx)
 	if err != nil {
 		return nil, trace.Wrap(err)
 	}
 
-	return filterDatabasesByLabels(rdsDatabases, f.cfg.Labels, f.log), nil
+	return filterDatabasesByLabels(rdsDatabases, f.cfg.Labels, f.log).AsResources(), nil
 }
 
 // getRDSDatabases returns a list of database resources representing RDS instances.
 func (f *rdsDBInstancesFetcher) getRDSDatabases(ctx context.Context) (types.Databases, error) {
-	instances, err := getAllDBInstances(ctx, f.cfg.RDS, common.MaxPages, f.log)
+	instances, err := getAllDBInstances(ctx, f.cfg.RDS, maxAWSPages, f.log)
 	if err != nil {
-		return nil, common.ConvertError(err)
+		return nil, trace.Wrap(libcloudaws.ConvertRequestFailureError(err))
 	}
 	databases := make(types.Databases, 0, len(instances))
 	for _, instance := range instances {
@@ -151,12 +155,14 @@ func (f *rdsDBInstancesFetcher) String() string {
 
 // rdsAuroraClustersFetcher retrieves RDS Aurora clusters.
 type rdsAuroraClustersFetcher struct {
-	cfg rdsFetcherConfig
+	awsFetcher
+
+	cfg RDSFetcherConfig
 	log logrus.FieldLogger
 }
 
-// newRDSAuroraClustersFetcher returns a new RDS Aurora fetcher instance.
-func newRDSAuroraClustersFetcher(config rdsFetcherConfig) (Fetcher, error) {
+// NewRDSAuroraClustersFetcher returns a new RDS Aurora fetcher instance.
+func NewRDSAuroraClustersFetcher(config RDSFetcherConfig) (common.Fetcher, error) {
 	if err := config.CheckAndSetDefaults(); err != nil {
 		return nil, trace.Wrap(err)
 	}
@@ -171,20 +177,20 @@ func newRDSAuroraClustersFetcher(config rdsFetcherConfig) (Fetcher, error) {
 }
 
 // Get returns Aurora clusters matching the watcher's selectors.
-func (f *rdsAuroraClustersFetcher) Get(ctx context.Context) (types.Databases, error) {
+func (f *rdsAuroraClustersFetcher) Get(ctx context.Context) (types.ResourcesWithLabels, error) {
 	auroraDatabases, err := f.getAuroraDatabases(ctx)
 	if err != nil {
 		return nil, trace.Wrap(err)
 	}
 
-	return filterDatabasesByLabels(auroraDatabases, f.cfg.Labels, f.log), nil
+	return filterDatabasesByLabels(auroraDatabases, f.cfg.Labels, f.log).AsResources(), nil
 }
 
 // getAuroraDatabases returns a list of database resources representing RDS clusters.
 func (f *rdsAuroraClustersFetcher) getAuroraDatabases(ctx context.Context) (types.Databases, error) {
-	clusters, err := getAllDBClusters(ctx, f.cfg.RDS, common.MaxPages, f.log)
+	clusters, err := getAllDBClusters(ctx, f.cfg.RDS, maxAWSPages, f.log)
 	if err != nil {
-		return nil, common.ConvertError(err)
+		return nil, trace.Wrap(libcloudaws.ConvertRequestFailureError(err))
 	}
 	databases := make(types.Databases, 0, len(clusters))
 	for _, cluster := range clusters {
@@ -325,7 +331,7 @@ func retryWithIndividualEngineFilters(log logrus.FieldLogger, engines []string, 
 	if err == nil {
 		return nil
 	}
-	if !common.IsUnrecognizedAWSEngineNameError(err) {
+	if !isUnrecognizedAWSEngineNameError(err) {
 		return trace.Wrap(err)
 	}
 	log.WithError(err).Warn("Teleport supports an engine which is unrecognized in this AWS region. Querying engines individually.")
@@ -334,10 +340,19 @@ func retryWithIndividualEngineFilters(log logrus.FieldLogger, engines []string, 
 		if err == nil {
 			continue
 		}
-		if !common.IsUnrecognizedAWSEngineNameError(err) {
+		if !isUnrecognizedAWSEngineNameError(err) {
 			return trace.Wrap(err)
 		}
 		// skip logging unrecognized engine name error here, we already logged it in the initial attempt.
 	}
 	return nil
+}
+
+// isUnrecognizedAWSEngineNameError checks if the err is non-nil and came from using an engine filter that the
+// AWS region does not recognize.
+func isUnrecognizedAWSEngineNameError(err error) bool {
+	if err == nil {
+		return false
+	}
+	return strings.Contains(strings.ToLower(err.Error()), "unrecognized engine name")
 }
