+++ lib/srv/discovery/fetchers/gke.go
@@ -18,6 +18,7 @@ package fetchers
 
 import (
 	"context"
+	"fmt"
 
 	containerpb "cloud.google.com/go/container/apiv1/containerpb"
 	"github.com/gravitational/trace"
@@ -26,6 +27,7 @@ import (
 	"github.com/gravitational/teleport/api/types"
 	"github.com/gravitational/teleport/lib/cloud/gcp"
 	"github.com/gravitational/teleport/lib/services"
+	"github.com/gravitational/teleport/lib/srv/discovery/common"
 )
 
 // GKEFetcherConfig configures the GKE fetcher.
@@ -68,7 +70,7 @@ type gkeFetcher struct {
 }
 
 // NewGKEFetcher creates a new GKE fetcher configuration.
-func NewGKEFetcher(cfg GKEFetcherConfig) (Fetcher, error) {
+func NewGKEFetcher(cfg GKEFetcherConfig) (common.Fetcher, error) {
 	if err := cfg.CheckAndSetDefaults(); err != nil {
 		return nil, trace.Wrap(err)
 	}
@@ -112,6 +114,11 @@ func (a *gkeFetcher) ResourceType() string {
 
 func (a *gkeFetcher) Cloud() string {
 	return types.CloudGCP
+}
+
+func (a *gkeFetcher) String() string {
+	return fmt.Sprintf("gkeFetcher(ProjectID=%v, Location=%v, Labels=%v)",
+		a.ProjectID, a.Location, a.FilterLabels)
 }
 
 // gcpLabelsToTeleportLabels converts GKE labels to a labels map.
