+++ lib/srv/discovery/kube_watcher.go
@@ -19,17 +19,12 @@ package discovery
 import (
 	"context"
 	"sync"
-	"time"
 
 	"github.com/gravitational/trace"
-	"golang.org/x/sync/errgroup"
 
 	"github.com/gravitational/teleport/api/types"
 	"github.com/gravitational/teleport/lib/services"
-)
-
-const (
-	concurrencyLimit = 5
+	"github.com/gravitational/teleport/lib/srv/discovery/common"
 )
 
 func (s *Server) startKubeWatchers() error {
@@ -41,7 +36,7 @@ func (s *Server) startKubeWatchers() error {
 		mu            sync.Mutex
 	)
 
-	watcher, err := services.NewReconciler(
+	reconciler, err := services.NewReconciler(
 		services.ReconcilerConfig{
 			Matcher: func(_ types.ResourceWithLabels) bool { return true },
 			GetCurrentResources: func() types.ResourcesWithLabelsMap {
@@ -67,7 +62,7 @@ func (s *Server) startKubeWatchers() error {
 				defer mu.Unlock()
 				return kubeResources.ToMap()
 			},
-			Log:      s.Log,
+			Log:      s.Log.WithField("kind", types.KindKubernetesCluster),
 			OnCreate: s.onKubeCreate,
 			OnUpdate: s.onKubeUpdate,
 			OnDelete: s.onKubeDelete,
@@ -77,55 +72,33 @@ func (s *Server) startKubeWatchers() error {
 		return trace.Wrap(err)
 	}
 
+	watcher, err := common.NewWatcher(s.ctx, common.WatcherConfig{
+		Fetchers: s.kubeFetchers,
+		Log:      s.Log.WithField("kind", types.KindKubernetesCluster),
+	})
+	if err != nil {
+		return trace.Wrap(err)
+	}
+	go watcher.Start()
+
 	go func() {
-		t := time.NewTicker(5 * time.Minute)
-		defer t.Stop()
 		for {
-			newResources := s.fetchFetchersResources()
-			mu.Lock()
-			kubeResources = newResources
-			mu.Unlock()
-
-			if err := watcher.Reconcile(s.ctx); err != nil {
-				s.Log.WithError(err).Warn("Unable to reconcile resources.")
-			}
-
 			select {
-			case <-t.C:
+			case newResources := <-watcher.ResourcesC():
+				mu.Lock()
+				kubeResources = newResources
+				mu.Unlock()
+
+				if err := reconciler.Reconcile(s.ctx); err != nil {
+					s.Log.WithError(err).Warn("Unable to reconcile resources.")
+				}
+
 			case <-s.ctx.Done():
 				return
 			}
 		}
 	}()
 	return nil
-}
-
-func (s *Server) fetchFetchersResources() types.ResourcesWithLabels {
-	var (
-		newFetcherResources = make(types.ResourcesWithLabels, 0, 50)
-		fetchersLock        sync.Mutex
-		group, groupCtx     = errgroup.WithContext(s.ctx)
-	)
-	group.SetLimit(concurrencyLimit)
-	for _, fetcher := range s.kubeFetchers {
-		lFetcher := fetcher
-
-		group.Go(func() error {
-			resources, err := lFetcher.Get(groupCtx)
-			if err != nil {
-				s.Log.WithError(err).Warnf("Unable to fetch resources for %s at %s.", lFetcher.ResourceType(), lFetcher.Cloud())
-				// never return the error otherwise it will impact other watchers.
-				return nil
-			}
-			fetchersLock.Lock()
-			newFetcherResources = append(newFetcherResources, resources...)
-			fetchersLock.Unlock()
-			return nil
-		})
-	}
-	// error is discarded because we must run all fetchers until the end.
-	_ = group.Wait()
-	return newFetcherResources
 }
 
 func (s *Server) onKubeCreate(ctx context.Context, rwl types.ResourceWithLabels) error {
